ai-service/requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
pypdf
openai
python-dotenv
google-generativeai
